init1

+++ src/main/java/cn/asghoul/pojo/Address.java
@@ -1,31 +1,33 @@
 package cn.asghoul.pojo;
 
+import java.math.BigInteger;
+
 /*
  * 地址表：adress
  */
 public class Address {
 
 
-    private long addressId;//地址id
-    private long clientId;//用户id
+    private BigInteger addressId;//地址id
+    private BigInteger clientId;//用户id
     private String address;//地址
     private String telphone;//收货人电话
     private int defaultAddress;//默认地址状态
     private String name;//收货人姓名
 
-    public long getAddressId() {
+    public BigInteger getAddressId() {
         return addressId;
     }
 
-    public void setAddressId(long addressId) {
+    public void setAddressId(BigInteger addressId) {
         this.addressId = addressId;
     }
 
-    public long getClientId() {
+    public BigInteger getClientId() {
         return clientId;
     }
 
-    public void setClientId(long clientId) {
+    public void setClientId(BigInteger clientId) {
         this.clientId = clientId;
     }
